Added API::V2::BaseSerializer
